+++ src/main/resources/db/migration/V2017.01.12.17.22__create_exam_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE `exam` (
-  `id` tinyint NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(64) NOT NULL,
   `logic_q_count` int NOT NULL,
   `prog_q_count` int NOT NULL,
-  `status_id` tinyint NOT NULL,
+  `status_id` bigint NOT NULL,
   `time_created` datetime DEFAULT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT fk_exam_status FOREIGN KEY (status_id) REFERENCES exam_status(id)
