+++ src/main/resources/db/migration/V2017.01.12.17.49__create_user_exam_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `user_exam` (
-  `id` tinyint NOT NULL AUTO_INCREMENT,
-  `user_id` tinyint NOT NULL,
-  `exam_id` tinyint NOT NULL,
+  `id` bigint NOT NULL AUTO_INCREMENT,
+  `user_id` bigint NOT NULL,
+  `exam_id` bigint NOT NULL,
   `score` int NOT NULL,
   `time_created` datetime DEFAULT NULL,
   PRIMARY KEY (`id`),
